refactor: Remove Tom's Vulkan kernel
