--- crates/kernels/vulkan-and-cpu/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "kernel"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["dylib", "lib"]
-
-[dependencies]
-spirv-std = { git = "https://github.com/Rust-GPU/rust-gpu", rev = "b3eda4df9814b6176d3c0844eb67f70a22ebb378" }
-
-# Dependencies for GPU code
-[target.'cfg(target_arch = "spirv")'.dependencies]
-glam = { version = "0.30.8", default-features = false, features = ["libm"] }
-
-# Dependencies for CPU code
-[target.'cfg(not(target_arch = "spirv"))'.dependencies]
-glam = { version = "0.30.8", features = ["std", "bytemuck"] }
-bytemuck = "1.22.0"
-
-[dev-dependencies]
-googletest = "0.14.2"
-
-[package.metadata.rust-gpu.build]
-# Where to output the compiled shader. Defaults to where `cargo gpu` is called from.
-# TODO: Should it default to the root of the shader crate?
-output-dir = "./"
-# Set shader crate's cargo default-features
-default_features = true
-# We don't want the manifest.json so hide it in an ignored path
-manifest-file = "../../../target/shader-manifest.json"
-# The SPIRV compile target.
-target = "spirv-unknown-vulkan1.4"
-
-# Needed to save the un-rotated elevation data in a storage buffer.
-capabilities = [ "Int16" ]
-
-[lints]
-workspace = true
-
