--- crates/kernels/vulkan-and-cpu/src/elevations.rs
@@ -1,108 +1,0 @@
-//! Code for managing elevation data.
-
-/// The nodata value from NASA's SRTM data.
-pub const NODATA: f32 = -32768.0;
-
-#[cfg_attr(not(target_arch = "spirv"), derive(Debug, PartialEq, Eq))]
-#[expect(
-    clippy::exhaustive_enums,
-    reason = "There's never going to be more directions"
-)]
-/// The direction of a band from the observer's point of view. Whether it points North or South is
-/// not relevant, they're just opposite to each other.
-pub enum Direction {
-    /// A line of sight facing forward from the observer.
-    Forward,
-    /// A line of sight facing behind the observer.
-    Backward,
-}
-
-/// Manage fetching valid elevation data.
-pub struct Elevations {
-    /// Whether going forwards or backwards along the line of sight.
-    pub direction: Direction,
-    /// The Chocolate Box ID of the current elevation point in rotated space.
-    chocolate_box_id: usize,
-    /// The elevation at the observer.
-    pov_elevation: f32,
-    /// A record of the last valid elevation. Used to fill "nodata" regions.
-    ///
-    /// Note that even though most, if not all, of these "nodata" regions occur at sea, that
-    /// doesn't necessarily mean that a default elevation of 0 is best. There are various reasons
-    /// why the sea isn't always at a perfect 0 elevation, which we won't go into here. The
-    /// point being that we need to smoothly transition into "nodata" regions to avoid visual
-    /// artefacts.
-    last_valid_elevation: f32,
-}
-
-impl Elevations {
-    #[inline]
-    #[must_use]
-    /// Instantiate.
-    pub fn new(
-        elevations: &[f32],
-        direction: Direction,
-        chocolate_box_id: usize,
-        observer_height: f32,
-    ) -> Self {
-        let elevation = elevations[chocolate_box_id];
-        let last_valid_elevation = if Self::is_valid(elevation) {
-            elevation
-        } else {
-            0.0
-        };
-        let pov_elevation = last_valid_elevation + observer_height;
-
-        Self {
-            direction,
-            chocolate_box_id,
-            pov_elevation,
-            last_valid_elevation,
-        }
-    }
-
-    #[inline]
-    /// Get the next elevation.
-    pub fn next(&mut self, elevations: &[f32]) -> f32 {
-        match self.direction {
-            Direction::Forward => self.chocolate_box_id += 1,
-            Direction::Backward => self.chocolate_box_id -= 1,
-        }
-
-        let elevation = self.current_elevation(elevations);
-
-        if Self::is_valid(elevation) {
-            self.last_valid_elevation = elevation;
-            elevation - self.pov_elevation
-        } else {
-            self.last_valid_elevation - self.pov_elevation
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    /// Return the current elevation.
-    pub fn current_elevation(&self, elevations: &[f32]) -> f32 {
-        elevations[self.chocolate_box_id]
-    }
-
-    #[inline]
-    #[must_use]
-    /// Are we moving forward through the elevations?
-    pub const fn is_forward(&self) -> bool {
-        matches!(self.direction, crate::elevations::Direction::Forward)
-    }
-
-    #[inline]
-    #[must_use]
-    /// Are we moving backward through the elevations?
-    pub const fn is_backward(&self) -> bool {
-        matches!(self.direction, crate::elevations::Direction::Backward)
-    }
-
-    /// Is the elevation valid.
-    fn is_valid(elevation: f32) -> bool {
-        let is_nodata = elevation <= NODATA;
-        !elevation.is_nan() && !is_nodata
-    }
-}
