+++ crates/total-viewsheds/Cargo.toml
@@ -9,20 +9,16 @@ rust-version = "1.89"
 bytemuck = "1.22.0"
 clap = { version = "4.5.42", features = ["derive"] }
 color-eyre = "0.6.5"
-fjall = "2.11.2"
 gdal = "0.19.0"
 geo = "0.30.0"
 geojson = "0.24.2"
 image = { version = "0.25.6", default-features = false, features = ["png"] }
 itertools = "0.14.0"
-kernel = { path = "../kernels/vulkan-and-cpu" }
-pollster = "0.4"
 proj4rs = { version = "0.1.8", features = ["aeqd"] }
 serde = "1.0.219"
 serde_json = "1.0.143"
 tracing = { version = "0.1.41" }
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
-wgpu = { version = "27", default-features = false, features = ["spirv", "vulkan"] }
 rusqlite = { version = "0.38.0", features = ["bundled"] }
 memchr = "2.8.0"
 rayon = "1.12.0"
