+++ crates/total-viewsheds/src/compute/los.rs
