+++ crates/total-viewsheds/src/compute/rotation.rs
