+++ crates/total-viewsheds/src/compute/rotator.rs
