+++ crates/total-viewsheds/src/main.rs
@@ -28,15 +28,12 @@ use color_eyre::eyre::Result;
 use std::mem;
 use tracing_subscriber::{Layer as _, layer::SubscriberExt as _, util::SubscriberInitExt as _};
 
-/// Handling the running of computations.
-mod run {
-    pub mod compute;
-    pub mod parallel;
-    pub mod serial;
-}
 mod config;
+mod run;
+mod workers;
+
 /// cpu implements a CPU kernel for the longest line of sight
-mod cpu {
+mod compute {
     pub mod area_of_interest;
 
     /// los contains all the traits necessary for implementing a line of sight algorithm
@@ -44,15 +41,6 @@ mod cpu {
 
     mod rotation;
     mod rotator;
-
-    /// Database for viewsheds
-    pub mod storage {
-        pub mod db;
-        pub mod engine;
-        pub mod metadata;
-        pub mod segments;
-        pub mod worker;
-    }
 
     /// kernel is the exported kernel module
     pub mod kernel;
@@ -69,14 +57,21 @@ mod dem;
 mod dump_usage;
 mod los_pack;
 mod post_process;
+/// Database for viewsheds
+mod storage {
+    pub mod db;
+    pub mod engine;
+    pub mod metadata;
+    pub mod segments;
+    pub mod worker;
+}
 mod tile;
-mod vulkan;
+
 /// Various ways to output data.
 mod output {
     pub mod ascii;
     pub mod bresenham;
     pub mod png;
-    pub mod ring_data;
     pub mod tiff;
     pub mod viewshed;
 }
@@ -97,14 +92,10 @@ fn main() -> Result<()> {
                     geo::coord! {x: f64::from(coordinate.0), y: f64::from(coordinate.1)},
                 );
 
-                let is_sqlite = rusqlite::Connection::open(viewshed_config.db_path.clone()).is_ok();
-                let source = if is_sqlite {
-                    output::ring_data::Source::SQLite(viewshed_config.db_path.clone())
-                } else {
-                    output::ring_data::Source::Directory(viewshed_config.db_path.clone())
-                };
-
-                let viewshed = crate::output::viewshed::Viewshed::reconstruct(&source, geo_coord)?;
+                let viewshed = crate::output::viewshed::Viewshed::reconstruct(
+                    viewshed_config.db_path.clone(),
+                    geo_coord,
+                )?;
                 crate::output::viewshed::Reconstructor::save(
                     viewshed,
                     &viewshed_config.output_dir,
@@ -154,33 +145,19 @@ fn compute(config: &config::Compute) -> Result<()> {
 
     tracing::debug!("Created DEM: {dem:?}");
 
-    #[expect(
-        clippy::as_conversions,
-        clippy::cast_sign_loss,
-        clippy::cast_possible_truncation,
-        clippy::cast_precision_loss,
-        reason = "Sign loss and truncation aren't relevant"
-    )]
-    let max_line_of_sight_as_metres = (max_line_of_sight_as_points as f32 * tile.scale) as u32;
-
     // Free up RAM
     drop(tile);
 
-    let dem_metadata = crate::cpu::storage::metadata::MetaData {
+    let dem_metadata = crate::storage::metadata::MetaData {
         width: dem.width,
         scale: dem.scale,
         max_line_of_sight: max_line_of_sight_as_points,
-        reserved_ring_size: run::compute::Compute::ring_count_per_band(
-            config.rings_per_km,
-            max_line_of_sight_as_metres,
-        ),
         centre: dem.centre,
     };
 
     tracing::info!("Starting computations");
-    let compute_config = run::compute::Config {
+    let compute_config = run::Config {
         observer_height: config.observer_height,
-        backend: config.backend.clone(),
         process: config.process.clone(),
         output_directory: Some(config.output_dir.clone()),
         heatmap: config.heatmap,
@@ -188,7 +165,7 @@ fn compute(config: &config::Compute) -> Result<()> {
         thread_count: config.thread_count,
         disable_render_image: config.disable_image_render,
         viewsheds_db_path: config.viewsheds_db_path.clone(),
-        area_of_interest: crate::cpu::area_of_interest::Pruner::lonlat_coords_to_polygon(
+        area_of_interest: crate::compute::area_of_interest::Pruner::lonlat_coords_to_polygon(
             config.aoi_point.clone(),
             &dem_metadata,
         )?,
@@ -196,7 +173,12 @@ fn compute(config: &config::Compute) -> Result<()> {
         database_per_thread: config.database_per_thread,
     };
 
-    let mut compute = run::compute::Compute::new(compute_config, &mut dem)?;
+    let mut compute = run::Compute::new(compute_config, &mut dem)?;
     compute.run()?;
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    pub mod fixtures;
 }
