+++ crates/total-viewsheds/src/output/viewshed.rs
@@ -1,7 +1,7 @@
 //! Reconstruct _individual_ viewsheds, not total viewsheds.
 
-use color_eyre::eyre::{ContextCompat as _, Result};
-use geo::{BooleanOps as _, HasDimensions as _};
+use color_eyre::eyre::Result;
+use geo::BooleanOps as _;
 
 /// A viewshed-based coordinate is projected to a metric system where the anchor is the viewshed's
 /// point of view. The other option would be a metric projection with an anchor in the DEM centre,
@@ -21,33 +21,22 @@ pub struct Viewshed<'viewshed> {
 impl Viewshed<'_> {
     /// Reconstruct a viewshed.
     pub fn reconstruct(
-        source: &super::ring_data::Source,
+        db_path: std::path::PathBuf,
         pov_coord_lonlat: crate::projection::LonLatCoord,
     ) -> Result<geo::MultiPolygon> {
-        let ring_data = match source {
-            crate::output::ring_data::Source::Directory(directory) => {
-                &super::ring_data::AllData::new_from_fjall(directory)?
-            }
-            crate::output::ring_data::Source::SQLite(db_path) => {
-                &super::ring_data::AllData::new_from_sqlite(db_path)?
-            }
-            #[cfg(test)]
-            crate::output::ring_data::Source::RAM(data) => data,
-        };
-        tracing::debug!("Using metadata for ring data: {:?}", ring_data.metadata);
+        let db = crate::storage::db::DB::new(db_path)?;
+        let metadata = db.load_metadata()?;
+        tracing::debug!("Using metadata for ring data: {:?}", metadata);
 
-        let mut polygon = geo::MultiPolygon::empty();
         let dem = crate::dem::DEM::new(
-            ring_data.metadata.centre,
-            ring_data.metadata.width,
-            ring_data.metadata.scale,
-            ring_data.metadata.max_line_of_sight,
+            metadata.centre,
+            metadata.width,
+            metadata.scale,
+            metadata.max_line_of_sight,
         )?;
 
-        let pov_dem_coord = crate::projection::Converter::lonlat_to_dem_coord(
-            &ring_data.metadata,
-            pov_coord_lonlat,
-        )?;
+        let pov_dem_coord =
+            crate::projection::Converter::lonlat_to_dem_coord(&metadata, pov_coord_lonlat)?;
         tracing::info!(
             "Reconstructing viewshed for DEM-relative coord: {:?}.",
             pov_dem_coord
@@ -58,31 +47,10 @@ impl Viewshed<'_> {
             pov_coord: pov_dem_coord,
         };
 
-        let constructor_by_sectors = || -> Result<geo::MultiPolygon> {
-            for angle_integer in 0..crate::run::compute::SECTOR_STEPS {
-                let angle = f32::from(angle_integer);
-                let mut reconstructor =
-                    Reconstructor::new(&viewshed, ring_data.metadata.reserved_ring_size, angle)?;
-                reconstructor.sector_ring_data = ring_data.get_sector(angle_integer)?;
-                polygon = reconstructor.reconstruct_sector(polygon)?;
-            }
+        let mut reconstructor = Reconstructor::new(&viewshed, 0.0)?;
 
-            Ok(polygon)
-        };
-
-        polygon = match source {
-            crate::output::ring_data::Source::Directory(_) => constructor_by_sectors()?,
-            crate::output::ring_data::Source::SQLite(db_path) => {
-                let db = crate::cpu::storage::db::DB::new(db_path)?;
-                let mut reconstructor =
-                    Reconstructor::new(&viewshed, ring_data.metadata.reserved_ring_size, 0.0)?;
-
-                let segments = db.load_segments_for_tvs_id(reconstructor.pov_id)?;
-                reconstructor.parse_polar_segments(&segments)
-            }
-            #[cfg(test)]
-            crate::output::ring_data::Source::RAM(_) => constructor_by_sectors()?,
-        };
+        let segments = db.load_segments_for_tvs_id(reconstructor.pov_id)?;
+        let polygon = reconstructor.parse_polar_segments(&segments);
 
         Ok(polygon)
     }
@@ -117,12 +85,6 @@ impl Viewshed<'_> {
 pub struct Reconstructor<'viewshed> {
     /// Data for the entire viewshed.
     viewshed: &'viewshed Viewshed<'viewshed>,
-    /// Data about the visibility regions of each computed band of sight.
-    sector_ring_data: Vec<u32>,
-    /// Where we're currently reading sector data from.
-    cursor: usize,
-    /// Amount of reserved ring data space.
-    reserved_ring_size: usize,
     /// The DEM id of the observer.
     pov_id: u32,
     /// The current sector angle
@@ -135,17 +97,10 @@ impl<'viewshed> Reconstructor<'viewshed> {
     /// The reason we only reconstruct for a single angle is that the sector data (for the angle)
     /// is most useful exposed as an iterator. And it's not so easy with lifetimes to keep around
     /// the raw data and an iterator for each angle.
-    fn new(
-        viewshed: &'viewshed Viewshed<'viewshed>,
-        reserved_ring_size: usize,
-        angle: f32,
-    ) -> Result<Self> {
+    fn new(viewshed: &'viewshed Viewshed<'viewshed>, angle: f32) -> Result<Self> {
         let pov_id = viewshed.dem.dem_coord_to_id(viewshed.pov_coord);
         let reconstructor = Self {
             viewshed,
-            sector_ring_data: Vec::default(),
-            cursor: 0,
-            reserved_ring_size,
             pov_id,
             current_angle: angle,
         };
@@ -160,91 +115,10 @@ impl<'viewshed> Reconstructor<'viewshed> {
         Ok(reconstructor)
     }
 
-    /// Extract and reconstruct a single viewshed from all the ring data for all possible viewsheds.
-    pub fn reconstruct_sector(&mut self, viewshed: geo::MultiPolygon) -> Result<geo::MultiPolygon> {
-        tracing::debug!(
-            "Building viewshed for sector {} using ring data of length {}",
-            self.current_angle,
-            self.sector_ring_data.len()
-        );
-        self.parse_fjall_sector(viewshed)
-    }
-
-    /// Read the next value in the ring data array.
-    fn read_next_value(&mut self) -> Result<u32> {
-        let value = *self
-            .sector_ring_data
-            .get(self.cursor)
-            .context("Couldn't get next ring in ring data")?;
-        self.cursor += 1;
-        Ok(value)
-    }
-
-    /// Parse an entire sector (angle) of ring data.
-    fn parse_fjall_sector(
-        &mut self,
-        mut viewshed_so_far: geo::MultiPolygon,
-    ) -> Result<geo::MultiPolygon> {
-        let tvs_id = self.viewshed.dem.pov_id_to_tvs_id(u64::from(self.pov_id));
-        let rotated_tvs_id = kernel::rotation::Rotator::rotate_index(
-            u32::try_from(tvs_id)?,
-            self.viewshed.dem.tvs_width,
-            self.current_angle,
-        );
-
-        self.cursor = rotated_tvs_id * self.reserved_ring_size;
-
-        let computable_points = usize::try_from(self.viewshed.dem.computable_points_count)?;
-        let wrap_to_backwards_ids = (rotated_tvs_id + computable_points) * self.reserved_ring_size;
-
-        let max_rings = u32::try_from((self.reserved_ring_size - 2).div_euclid(2))?;
-        for direction in [
-            kernel::elevations::Direction::Forward,
-            kernel::elevations::Direction::Backward,
-        ] {
-            // We divide by 2 because every ring must have both an opening and a closing.
-            let mut no_of_ring_values = self.read_next_value()?.div_euclid(2);
-
-            if no_of_ring_values == 0 {
-                // These are most common outside the circle area of the TVS region.
-                self.cursor = wrap_to_backwards_ids;
-                continue;
-            }
-            if no_of_ring_values > max_rings {
-                tracing::warn!(
-                    "More rings in band than reserved rings ({} > {}) for point {:?}",
-                    no_of_ring_values,
-                    max_rings,
-                    direction
-                );
-                no_of_ring_values = max_rings;
-            }
-
-            for ring in 0..no_of_ring_values {
-                let opening = if ring == 0 {
-                    0
-                } else {
-                    self.read_next_value()?
-                };
-                let closing = self.read_next_value()?;
-
-                let polygon = self.make_visible_polygon(opening, closing, &direction);
-                viewshed_so_far = viewshed_so_far.union(&polygon);
-                if viewshed_so_far.is_empty() {
-                    color_eyre::eyre::bail!("Invalid polygon: {polygon:?}");
-                }
-            }
-
-            self.cursor = wrap_to_backwards_ids;
-        }
-
-        Ok(viewshed_so_far)
-    }
-
     /// Convert polar segements to `GeoJson`.
     pub fn parse_polar_segments(
         &mut self,
-        data: &[Vec<crate::cpu::storage::segments::Segment>],
+        data: &[Vec<crate::storage::segments::Segment>],
     ) -> geo::MultiPolygon {
         let mut viewshed_so_far = geo::MultiPolygon::empty();
         for (angle, segments) in data.iter().enumerate() {
@@ -260,11 +134,7 @@ impl<'viewshed> Reconstructor<'viewshed> {
             for segment in segments {
                 let opening = u32::from(segment.start());
                 let closing = u32::from(segment.start() + segment.distance());
-                let polygon = self.make_visible_polygon(
-                    opening,
-                    closing,
-                    &kernel::elevations::Direction::Forward,
-                );
+                let polygon = self.make_visible_polygon(opening, closing);
                 viewshed_so_far = viewshed_so_far.union(&polygon);
             }
         }
@@ -273,20 +143,13 @@ impl<'viewshed> Reconstructor<'viewshed> {
     }
 
     /// Convert an index along a line of sight into a coordinate.
-    fn index_to_coordinate(
-        &self,
-        index: u32,
-        direction: &kernel::elevations::Direction,
-    ) -> Coordinate {
-        let angle = match direction {
-            kernel::elevations::Direction::Forward => self.current_angle.to_radians(),
-            kernel::elevations::Direction::Backward => (self.current_angle + 180.0).to_radians(),
-        };
+    fn index_to_coordinate(&self, index: u32) -> Coordinate {
+        let radians = self.current_angle.to_radians();
         let distance = f64::from(index);
 
         Coordinate(geo::coord! {
-            x: distance * f64::from(angle.cos()),
-            y: distance * f64::from(angle.sin())
+            x: distance * f64::from(radians.cos()),
+            y: distance * f64::from(radians.sin())
         })
     }
 
@@ -307,14 +170,9 @@ impl<'viewshed> Reconstructor<'viewshed> {
     }
 
     /// Make a single polygon representing a visible region of the planet.
-    fn make_visible_polygon(
-        &self,
-        opening_index: u32,
-        closing_index: u32,
-        direction: &kernel::elevations::Direction,
-    ) -> geo::Polygon {
-        let opening_coord = self.index_to_coordinate(opening_index, direction);
-        let closing_coord = self.index_to_coordinate(closing_index, direction);
+    fn make_visible_polygon(&self, opening_index: u32, closing_index: u32) -> geo::Polygon {
+        let opening_coord = self.index_to_coordinate(opening_index);
+        let closing_coord = self.index_to_coordinate(closing_index);
 
         let spread = 0.5001f64;
         let bottom_left = Self::rotate_by(opening_coord, spread);
@@ -399,14 +257,11 @@ impl<'viewshed> Reconstructor<'viewshed> {
 
 #[cfg(test)]
 mod test {
+    use super::*;
     use crate::{output::ascii::assert_viewshed, run};
 
-    use super::*;
-
-    const RESERVED_RING_SIZE: usize = crate::run::compute::Compute::ring_count_per_band(5000.0, 3);
-
     fn builder<'viewshed>(viewshed: &'viewshed Viewshed, angle: f32) -> Reconstructor<'viewshed> {
-        Reconstructor::new(viewshed, RESERVED_RING_SIZE, angle).unwrap()
+        Reconstructor::new(viewshed, angle).unwrap()
     }
 
     #[derive(Debug)]
@@ -418,23 +273,13 @@ mod test {
     }
 
     fn make_visible_polygon_for(setup: &VisiblePolygonFor) -> Vec<geo::Coord> {
-        let dem = crate::run::compute::test::make_dem(&kernel::tests::dems::single_peak_dem());
+        let dem = crate::run::test::make_dem(&crate::tests::fixtures::single_peak_dem());
         let viewshed = Viewshed {
             dem: &dem,
             pov_coord: crate::dem::Coordinate(setup.pov),
         };
-        let direction = if setup.angle < 180.0 {
-            kernel::elevations::Direction::Forward
-        } else {
-            kernel::elevations::Direction::Backward
-        };
-        let angle = match direction {
-            kernel::elevations::Direction::Forward => setup.angle,
-            kernel::elevations::Direction::Backward => setup.angle - 180.0,
-        };
-        let viewsheder = builder(&viewshed, angle);
-        let polygon =
-            viewsheder.make_visible_polygon(setup.opening_index, setup.closing_index, &direction);
+        let viewsheder = builder(&viewshed, setup.angle);
+        let polygon = viewsheder.make_visible_polygon(setup.opening_index, setup.closing_index);
 
         let mut polygon_as_dem_coords = Vec::new();
         for coord in &polygon.exterior().0 {
@@ -545,14 +390,15 @@ mod test {
         }
     }
 
-    fn viewshed_in_hole(backend: &crate::config::Backend) {
+    #[test]
+    fn viewshed_in_hole() {
         let temp_db = tempfile::NamedTempFile::new().unwrap();
         let viewshed = crate::output::ascii::make_viewshed(
-            &kernel::tests::dems::bigger_dem(),
+            &crate::tests::fixtures::bigger_dem(),
             geo::Coord { x: 5.0, y: 5.0 },
-            run::compute::Config {
-                dem_metadata: run::compute::test::big_dem_metadata(),
-                ..crate::run::compute::test::default_config(backend.clone(), &temp_db)
+            run::Config {
+                dem_metadata: run::test::big_dem_metadata(),
+                ..crate::run::test::default_config(&temp_db)
             },
         );
 
@@ -570,55 +416,47 @@ mod test {
             "████████████████████████",
             "████████████████████████",
         ];
-
-        match backend {
-            crate::config::Backend::VulkanCPU | crate::config::Backend::CPU => {
-                assert_viewshed(&viewshed, expected);
-            }
-            crate::config::Backend::Vulkan | crate::config::Backend::Cuda => {
-                panic!("We're not testing these.")
-            }
-        }
+        assert_viewshed(&viewshed, expected);
     }
 
-    fn viewshed_on_summit(backend: crate::config::Backend) {
+    #[test]
+    fn viewshed_on_summit() {
         let temp_db = tempfile::NamedTempFile::new().unwrap();
         let viewshed = crate::output::ascii::make_viewshed(
-            &kernel::tests::dems::bigger_dem(),
+            &crate::tests::fixtures::bigger_dem(),
             geo::Coord { x: 6.0, y: 6.0 },
-            run::compute::Config {
-                dem_metadata: run::compute::test::big_dem_metadata(),
-                ..crate::run::compute::test::default_config(backend, &temp_db)
+            run::Config {
+                dem_metadata: run::test::big_dem_metadata(),
+                ..crate::run::test::default_config(&temp_db)
             },
         );
 
-        assert_viewshed(
-            &viewshed,
-            &[
-                "████████████████████████",
-                "████████████████████████",
-                "███████▀▀ ▄▄▄▄▄ ▀▀██████",
-                "█████▀ ▄█████████▄ ▀████",
-                "████▀ █████████████ ▀███",
-                "████ ███████████████ ███",
-                "████ ███████████████ ███",
-                "████ ▀█████████████▀ ███",
-                "█████ ▀███████████▀ ████",
-                "██████▄ ▀▀█████▀▀ ▄█████",
-                "█████████▄▄▄▄▄▄▄████████",
-                "████████████████████████",
-            ],
-        );
+        let expected = [
+            "████████████████████████",
+            "████████████████████████",
+            "███████▀▀ ▄▄▄▄▄ ▀▀██████",
+            "█████▀ ▄█████████▄ ▀████",
+            "████▀ █████████████ ▀███",
+            "████ ███████████████ ███",
+            "████ ███████████████ ███",
+            "████ ▀█████████████▀ ███",
+            "█████ ▀███████████▀ ████",
+            "██████▄ ▀▀█████▀▀ ▄█████",
+            "█████████▄▄▄▄▄▄▄████████",
+            "████████████████████████",
+        ];
+        assert_viewshed(&viewshed, &expected);
     }
 
-    fn viewshed_near_summit(backend: &crate::config::Backend) {
+    #[test]
+    fn viewshed_near_summit() {
         let temp_db = tempfile::NamedTempFile::new().unwrap();
         let viewshed = crate::output::ascii::make_viewshed(
-            &kernel::tests::dems::bigger_dem(),
+            &crate::tests::fixtures::bigger_dem(),
             geo::Coord { x: 5.0, y: 6.0 },
-            run::compute::Config {
-                dem_metadata: run::compute::test::big_dem_metadata(),
-                ..crate::run::compute::test::default_config(backend.clone(), &temp_db)
+            run::Config {
+                dem_metadata: run::test::big_dem_metadata(),
+                ..crate::run::test::default_config(&temp_db)
             },
         );
 
@@ -636,89 +474,36 @@ mod test {
             "███████▄▄▄▄▄▄▄██████████",
             "████████████████████████",
         ];
-
-        match backend {
-            crate::config::Backend::VulkanCPU | crate::config::Backend::CPU => {
-                assert_viewshed(&viewshed, expected);
-            }
-            crate::config::Backend::Vulkan | crate::config::Backend::Cuda => {
-                panic!("We're not testing these.")
-            }
-        }
+        assert_viewshed(&viewshed, expected);
     }
 
-    fn viewshed_in_hole_tall_observer(backend: crate::config::Backend) {
+    #[test]
+    fn viewshed_in_hole_tall_observer() {
         let temp_db = tempfile::NamedTempFile::new().unwrap();
         let viewshed = crate::output::ascii::make_viewshed(
-            &kernel::tests::dems::bigger_dem(),
+            &crate::tests::fixtures::bigger_dem(),
             geo::Coord { x: 5.0, y: 5.0 },
-            run::compute::Config {
+            run::Config {
                 observer_height: 20.0,
-                dem_metadata: run::compute::test::big_dem_metadata(),
-                ..crate::run::compute::test::default_config(backend, &temp_db)
+                dem_metadata: run::test::big_dem_metadata(),
+                ..crate::run::test::default_config(&temp_db)
             },
         );
 
-        assert_viewshed(
-            &viewshed,
-            &[
-                "████████████████████████",
-                "█████▀▀ ▄▄▄▄▄ ▀▀████████",
-                "███▀ ▄█████████▄ ▀██████",
-                "██▀ █████████████ ▀█████",
-                "██ ███████████████ █████",
-                "██ ███████████████ █████",
-                "██ ▀█████████████▀ █████",
-                "███ ▀███████████▀ ██████",
-                "████▄ ▀▀█████▀▀ ▄███████",
-                "███████▄▄▄▄▄▄▄██████████",
-                "████████████████████████",
-                "████████████████████████",
-            ],
-        );
-    }
-
-    mod gpu {
-        #[test]
-        fn viewshed_in_hole() {
-            super::viewshed_in_hole(&crate::config::Backend::VulkanCPU);
-        }
-
-        #[test]
-        fn viewshed_on_summit() {
-            super::viewshed_on_summit(crate::config::Backend::VulkanCPU);
-        }
-
-        #[test]
-        fn viewshed_near_summit() {
-            super::viewshed_near_summit(&crate::config::Backend::VulkanCPU);
-        }
-
-        #[test]
-        fn viewshed_in_hole_tall_observer() {
-            super::viewshed_in_hole_tall_observer(crate::config::Backend::VulkanCPU);
-        }
-    }
-
-    mod cpu {
-        #[test]
-        fn viewshed_in_hole() {
-            super::viewshed_in_hole(&crate::config::Backend::CPU);
-        }
-
-        #[test]
-        fn viewshed_on_summit() {
-            super::viewshed_on_summit(crate::config::Backend::CPU);
-        }
-
-        #[test]
-        fn viewshed_near_summit() {
-            super::viewshed_near_summit(&crate::config::Backend::CPU);
-        }
-
-        #[test]
-        fn viewshed_in_hole_tall_observer() {
-            super::viewshed_in_hole_tall_observer(crate::config::Backend::CPU);
-        }
+        let expected = &[
+            "████████████████████████",
+            "█████▀▀ ▄▄▄▄▄ ▀▀████████",
+            "███▀ ▄█████████▄ ▀██████",
+            "██▀ █████████████ ▀█████",
+            "██ ███████████████ █████",
+            "██ ███████████████ █████",
+            "██ ▀█████████████▀ █████",
+            "███ ▀███████████▀ ██████",
+            "████▄ ▀▀█████▀▀ ▄███████",
+            "███████▄▄▄▄▄▄▄██████████",
+            "████████████████████████",
+            "████████████████████████",
+        ];
+        assert_viewshed(&viewshed, expected);
     }
 }
