--- crates/total-viewsheds/src/run/compute.rs
@@ -1,571 +1,0 @@
-//! The main entrypoint for running computations.
-
-use color_eyre::Result;
-use std::path::PathBuf;
-
-/// The number of angles we rotate through. The other half are done via "backwards" lines of sight.
-pub const SECTOR_STEPS: u16 = 180;
-
-/// Handles all the computations.
-pub struct Compute<'compute>
-where
-    'static: 'compute,
-{
-    /// User configuration.
-    pub config: Config,
-    /// Vulkan GPU manager
-    pub vulkan: Option<crate::vulkan::Vulkan>,
-    /// Storage interface for conputed ring (viewshed) data.
-    fjall_storage: Option<crate::output::ring_data::FjallStorage>,
-    /// The Digital Elevation Model that we're computing.
-    pub dem: &'compute mut crate::dem::DEM,
-    /// The constants for each kernel computation.
-    pub constants: kernel::constants::Constants,
-    /// The amount of reserved memory for ring data.
-    pub total_reserved_rings: usize,
-    /// Keeps track of the cumulative surfaces from every angle.
-    pub total_surfaces: Vec<f32>,
-    /// Keeps track of the ring (viewshed) data.
-    pub ring_data: Vec<Vec<u32>>,
-    /// Keeps track of the longest lines of sight.
-    pub longest_lines: Vec<crate::los_pack::LineOfSightPacked>,
-}
-
-#[derive(Clone)]
-/// Configuration for computing.
-pub struct Config {
-    /// The height of the observer that views viewsheds.
-    pub observer_height: f32,
-    /// Where to run the kernel computations
-    pub backend: crate::config::Backend,
-    /// What to compute.
-    pub process: Vec<crate::config::Process>,
-    /// Output directory
-    pub output_directory: Option<std::path::PathBuf>,
-    /// How to normalise the heatmap data.
-    pub heatmap: crate::config::HeatmapNormalisation,
-    /// Refractoin coefficient
-    pub refraction: f32,
-    /// Number of threads for computation
-    pub thread_count: usize,
-    /// Disables the rendering of PNG images (good for long runs)
-    pub disable_render_image: bool,
-    /// Where to store the viewshed
-    pub viewsheds_db_path: PathBuf,
-    /// Metadata about the DEM and compute run.
-    pub dem_metadata: crate::cpu::storage::metadata::MetaData,
-    /// Polygon for Area of Interest
-    pub area_of_interest: geo::Polygon,
-    /// Should a database aggregate per thread?
-    pub database_per_thread: bool,
-}
-
-impl<'compute> Compute<'compute> {
-    /// Instantiate.
-    pub fn new(config: Config, dem: &'compute mut crate::dem::DEM) -> Result<Self> {
-        let total_bands = dem.computable_points_count * 2;
-
-        let rings_per_band = if Self::is_process_viewsheds(&config.process) {
-            config.dem_metadata.reserved_ring_size
-        } else {
-            1
-        };
-        let total_reserved_rings = if Self::is_process_viewsheds(&config.process) {
-            usize::try_from(total_bands)? * rings_per_band
-        } else {
-            1
-        };
-
-        let fjall_storage = if Self::is_process_viewsheds(&config.process) {
-            match &config.output_directory {
-                Some(output_directory) => Some(crate::output::ring_data::FjallStorage::new(
-                    output_directory,
-                )?),
-                None => None,
-            }
-        } else {
-            None
-        };
-
-        if Self::is_process_viewsheds(&config.process) && !cfg!(any(test, feature = "ring_data")) {
-            color_eyre::eyre::bail!(
-                "Viewshed storage is only supported with the ring_data feature, \
-                please recompile with --features=ring_data"
-            );
-        }
-
-        let constants = kernel::constants::Constants {
-            total_bands,
-            max_los_as_points: dem.max_los_as_points,
-            dem_width: dem.width,
-            tvs_width: dem.tvs_width,
-            observer_height: config.observer_height,
-            reserved_rings_per_band: u32::try_from(rings_per_band)?,
-            process: Self::bitmask_flags_for_kernel(&config.process),
-            scale: config.dem_metadata.scale,
-            refraction: config.refraction,
-            ..Default::default()
-        };
-
-        // We only need the "chocolate box" section of rotations to do visibility calculations.
-        let rotations_size = kernel::chocolate_box::size(dem.width, dem.tvs_width);
-
-        let vulkan = if matches!(config.backend, crate::config::Backend::Vulkan) {
-            let elevations = dem.elevations.clone();
-            dem.elevations = Vec::new(); // Free up some RAM.
-            Some(crate::vulkan::Vulkan::new(
-                constants,
-                elevations,
-                usize::try_from(rotations_size)?,
-                total_reserved_rings,
-            )?)
-        } else {
-            None
-        };
-
-        Ok(Self {
-            config,
-            vulkan,
-            fjall_storage,
-            dem,
-            constants,
-            total_reserved_rings,
-            total_surfaces: Vec::default(),
-            ring_data: Vec::default(),
-            longest_lines: Vec::default(),
-        })
-    }
-
-    #[expect(
-        clippy::as_conversions,
-        clippy::cast_precision_loss,
-        clippy::cast_possible_truncation,
-        clippy::cast_sign_loss,
-        reason = "Accuracy isn't needed, we're just calculating a value to help find minimum RAM usage."
-    )]
-    /// Calculate the expected number of rings per band of sight.
-    pub const fn ring_count_per_band(rings_per_km: f32, max_line_of_sight: u32) -> usize {
-        let meters_per_km = 1000.0;
-        let band_length_in_km = (max_line_of_sight as f32) / meters_per_km;
-        (band_length_in_km * rings_per_km) as usize
-    }
-
-    /// Are we computing everything?
-    fn is_process_everything(process: &[crate::config::Process]) -> bool {
-        process.contains(&crate::config::Process::All)
-    }
-
-    /// Are we computing total surface areas?
-    pub fn is_process_surfaces(process: &[crate::config::Process]) -> bool {
-        Self::is_process_everything(process)
-            || process.contains(&crate::config::Process::TotalSurfaces)
-    }
-
-    /// Are we computing viewsheds?
-    pub fn is_process_viewsheds(process: &[crate::config::Process]) -> bool {
-        Self::is_process_everything(process) || process.contains(&crate::config::Process::Viewsheds)
-    }
-
-    /// Are we computing total surface areas?
-    pub fn is_process_longest_lines(process: &[crate::config::Process]) -> bool {
-        Self::is_process_everything(process)
-            || process.contains(&crate::config::Process::LongestLines)
-    }
-
-    /// Do all computations.
-    pub fn run(&mut self) -> Result<()> {
-        if Self::is_process_viewsheds(&self.config.process)
-            && self.config.output_directory.is_some()
-        {
-            self.save_ring_metadata()?;
-        }
-
-        if matches!(self.config.backend, crate::config::Backend::CPU) {
-            self.run_parallel()?;
-        } else {
-            self.run_sequential()?;
-        }
-
-        Ok(())
-    }
-
-    /// The metadata needed to reconstruct viewsheds based on the DEM and reserved rings.
-    pub fn metadata(&self) -> Result<crate::cpu::storage::metadata::MetaData> {
-        Ok(crate::cpu::storage::metadata::MetaData {
-            width: self.dem.width,
-            scale: self.dem.scale,
-            max_line_of_sight: self.dem.max_los_as_points,
-            reserved_ring_size: usize::try_from(self.constants.reserved_rings_per_band)?,
-            centre: self.dem.centre,
-        })
-    }
-
-    /// Save band deltas to cache.
-    pub fn save_sector_ring_data(&self, sector: u16, ring_data: &[u32]) -> Result<()> {
-        let Some(storage) = self.fjall_storage.as_ref() else {
-            color_eyre::eyre::bail!("Tried to save sector ring data without any active storage.");
-        };
-
-        storage.save_sector(sector, ring_data)?;
-        Ok(())
-    }
-
-    /// Save the metadata for the ring data (aka viewsheds).
-    pub fn save_ring_metadata(&self) -> Result<()> {
-        let Some(storage) = self.fjall_storage.as_ref() else {
-            color_eyre::eyre::bail!("Tried to save ring metadata without any active storage.");
-        };
-
-        storage.save_metadata(&self.metadata()?)?;
-        Ok(())
-    }
-
-    /// Render a heatmap and `.bt` file of the total surface areas for each point within the computable area of the
-    /// DEM.
-    pub fn render_total_surfaces(&self) -> Result<()> {
-        let Some(output_dir) = &self.config.output_directory else {
-            return Ok(());
-        };
-
-        crate::output::tiff::save(
-            self.dem,
-            &self.total_surfaces,
-            &output_dir.join("total_surfaces.tiff"),
-        )?;
-
-        if self.config.disable_render_image {
-            return Ok(());
-        }
-
-        crate::output::png::save(
-            &self.total_surfaces,
-            self.dem.tvs_width,
-            self.dem.tvs_width,
-            output_dir.join("total_surfaces.png"),
-            self.config.heatmap,
-        )?;
-
-        Ok(())
-    }
-
-    /// Render a heatmap and `.bt` of the longest lines of sight for each point within the computable area of the
-    /// DEM.
-    pub fn render_longest_lines(&self) -> Result<()> {
-        let Some(output_dir) = &self.config.output_directory else {
-            return Ok(());
-        };
-
-        let packed_lines = self
-            .longest_lines
-            .iter()
-            .map(|&packed| packed.as_f32())
-            .collect::<Vec<_>>();
-
-        crate::output::tiff::save(
-            self.dem,
-            &packed_lines,
-            &output_dir.join("longest_lines.tiff"),
-        )?;
-
-        if self.config.disable_render_image {
-            return Ok(());
-        }
-
-        let distances = self
-            .longest_lines
-            .iter()
-            .map(|los| {
-                #[expect(
-                    clippy::as_conversions,
-                    clippy::cast_precision_loss,
-                    reason = "Distances always fit in u32"
-                )]
-                {
-                    los.distance() as f32
-                }
-            })
-            .collect::<Vec<_>>();
-
-        crate::output::png::save(
-            &distances,
-            self.dem.tvs_width,
-            self.dem.tvs_width,
-            output_dir.join("longest_lines.png"),
-            self.config.heatmap,
-        )?;
-
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-pub mod test {
-    use super::*;
-    use googletest::prelude::*;
-
-    pub fn make_dem(elevations: &[i16]) -> crate::dem::DEM {
-        let width = elevations.len().isqrt() as u32;
-        let mut dem = crate::dem::DEM::new(
-            crate::projection::LonLatCoord((33.33, 33.33).into()),
-            width,
-            1.0,
-            width / 3,
-        )
-        .unwrap();
-        dem.elevations = elevations.into();
-        dem
-    }
-
-    pub fn default_metadata() -> crate::cpu::storage::metadata::MetaData {
-        crate::cpu::storage::metadata::MetaData {
-            reserved_ring_size: 100,
-            scale: 1.0,
-            ..Default::default()
-        }
-    }
-
-    pub fn big_dem_metadata() -> crate::cpu::storage::metadata::MetaData {
-        crate::cpu::storage::metadata::MetaData {
-            width: 12,
-            max_line_of_sight: 4,
-            centre: crate::projection::LonLatCoord((33.33, 33.33).into()),
-            ..crate::run::compute::test::default_metadata()
-        }
-    }
-
-    pub fn default_config(
-        backend: crate::config::Backend,
-        temp_db: &tempfile::NamedTempFile,
-    ) -> Config {
-        Config {
-            observer_height: 0.8,
-            backend,
-            process: vec![
-                crate::config::Process::TotalSurfaces,
-                crate::config::Process::Viewsheds,
-                crate::config::Process::LongestLines,
-            ],
-            output_directory: None,
-            heatmap: crate::config::HeatmapNormalisation::UnitScale,
-            refraction: 0.13f32,
-            thread_count: 1, // single thread it for consistency
-            disable_render_image: false,
-            viewsheds_db_path: temp_db.path().into(),
-            dem_metadata: default_metadata(),
-            area_of_interest: geo::Polygon::empty(),
-            database_per_thread: false,
-        }
-    }
-
-    pub fn compute(dem: &mut crate::dem::DEM, config: Config) -> Compute<'_> {
-        let mut compute = Compute::new(config, dem).unwrap();
-        compute.run().unwrap();
-        compute
-    }
-
-    fn total_surfaces(backend: crate::config::Backend) {
-        let mut dem = make_dem(&kernel::tests::dems::bigger_dem());
-        let temp_db = tempfile::NamedTempFile::new().unwrap();
-        let compute = compute(&mut dem, default_config(backend, &temp_db));
-        #[rustfmt::skip]
-        assert_eq!(
-            compute.total_surfaces,
-            [
-                0.0, 0.0,       0.0,       0.0,
-                0.0, 6.283163,  23.213009, 0.0,
-                0.0, 48.066483, 62.832096, 0.0,
-                0.0, 0.0,       0.0,       0.0
-            ]
-        );
-    }
-
-    #[expect(
-        clippy::as_conversions,
-        clippy::cast_precision_loss,
-        reason = "Distances always fit in u32"
-    )]
-    fn longest_lines(backend: &crate::config::Backend) {
-        let mut dem = make_dem(&kernel::tests::dems::bigger_dem());
-        let temp_db = tempfile::NamedTempFile::new().unwrap();
-        let config = Config {
-            dem_metadata: big_dem_metadata(),
-            ..crate::run::compute::test::default_config(backend.clone(), &temp_db)
-        };
-        let compute = compute(&mut dem, config);
-
-        #[rustfmt::skip]
-        expect_eq!(
-            compute.longest_lines.iter()
-            .map(|los| los.distance() as f32)
-            .collect::<Vec<_>>(),
-            [
-                0.0, 0.0, 0.0, 0.0,
-                0.0, 1.0, 4.0, 0.0,
-                0.0, 4.0, 4.0, 0.0,
-                0.0, 0.0, 0.0, 0.0
-            ]
-        );
-
-        #[rustfmt::skip]
-        let nearest_neighbour = [
-            0, 0,  0,  0,
-            0, 0,  12, 0,
-            0, 46, 0,  0,
-            0, 0,  0,  0
-        ];
-
-        #[rustfmt::skip]
-        let skew_rotation = [
-            0, 0,  0,  0,
-            0, 0, 119, 0,
-            0, 0,  0,  0,
-            0, 0,  0,  0
-        ];
-
-        expect_eq!(
-            compute
-                .longest_lines
-                .iter()
-                .map(|packed| packed.angle())
-                .collect::<Vec<_>>(),
-            if matches!(backend, crate::config::Backend::CPU) {
-                skew_rotation
-            } else {
-                nearest_neighbour
-            }
-        );
-    }
-
-    mod vulkan_cpu {
-        use googletest::prelude::*;
-
-        #[test]
-        fn total_surfaces() {
-            super::total_surfaces(crate::config::Backend::VulkanCPU);
-        }
-
-        #[gtest]
-        fn longest_lines() {
-            super::longest_lines(&crate::config::Backend::VulkanCPU);
-        }
-    }
-
-    mod cpu {
-        use super::{compute, make_dem};
-        use crate::config::Backend;
-        use googletest::prelude::*;
-
-        #[rustfmt::skip]
-        const EXPECTED_SURFACES: [f32; 16] = [
-            0.0, 0.0,      0.0,       0.0,
-            0.0, 6.283163, 21.118547,  0.0,
-            0.0, 47.90945, 62.832096, 0.0,
-            0.0, 0.0,      0.0,       0.0
-        ];
-
-        #[test]
-        fn total_surfaces() {
-            let mut dem = make_dem(&kernel::tests::dems::bigger_dem());
-            let temp_db = tempfile::NamedTempFile::new().unwrap();
-            let config = super::Config {
-                dem_metadata: crate::run::compute::test::big_dem_metadata(),
-                ..super::default_config(Backend::CPU, &temp_db)
-            };
-            let compute = compute(&mut dem, config);
-            assert_eq!(compute.total_surfaces, EXPECTED_SURFACES);
-        }
-
-        #[gtest]
-        fn longest_lines() {
-            super::longest_lines(&crate::config::Backend::CPU);
-        }
-
-        #[gtest]
-        fn refraction_affects_visibility() {
-            let mut dem_for_no_refraction = make_dem(&kernel::tests::dems::bigger_dem());
-            let temp_db_for_no_refraction = tempfile::NamedTempFile::new().unwrap();
-            let compute_no_refraction = super::compute(
-                &mut dem_for_no_refraction,
-                super::Config {
-                    // Our "bigger_dem" isn't actually big enough for a 0.0 refraction to have an
-                    // affect. We already test for default refraction above, so may as well test for
-                    // 0.0 refraction here just in case there's some unexpected divergence.
-                    refraction: 0.0,
-                    dem_metadata: crate::run::compute::test::big_dem_metadata(),
-                    ..super::default_config(Backend::CPU, &temp_db_for_no_refraction)
-                },
-            );
-            expect_eq!(compute_no_refraction.total_surfaces, EXPECTED_SURFACES);
-
-            let mut dem_for_very_refraction = make_dem(&kernel::tests::dems::bigger_dem());
-            let temp_db_for_very_refraction = tempfile::NamedTempFile::new().unwrap();
-            let compute_very_refraction = super::compute(
-                &mut dem_for_very_refraction,
-                super::Config {
-                    refraction: -kernel::kernel::EARTH_DIAMETER,
-                    dem_metadata: crate::run::compute::test::big_dem_metadata(),
-                    ..super::default_config(Backend::CPU, &temp_db_for_very_refraction)
-                },
-            );
-            #[rustfmt::skip]
-            expect_eq!(
-                compute_very_refraction.total_surfaces,
-                [
-                    0.0, 0.0,      0.0,       0.0,
-                    0.0, 6.283163, 9.424777,  0.0,
-                    0.0, 14.468757,38.083126, 0.0,
-                    0.0, 0.0,      0.0,       0.0
-                ]
-            );
-        }
-
-        #[gtest]
-        fn scale_affects_visibility() {
-            let mut dem_for_small_scale = make_dem(&kernel::tests::dems::bigger_dem());
-            let temp_db_for_small_scale = tempfile::NamedTempFile::new().unwrap();
-            let compute_small_scale = super::compute(
-                &mut dem_for_small_scale,
-                super::Config {
-                    dem_metadata: crate::cpu::storage::metadata::MetaData {
-                        scale: 0.01,
-                        ..crate::run::compute::test::big_dem_metadata()
-                    },
-                    ..super::default_config(Backend::CPU, &temp_db_for_small_scale)
-                },
-            );
-            #[rustfmt::skip]
-            expect_eq!(
-                compute_small_scale.total_surfaces,
-                [
-                    0.0, 0.0,       0.0,       0.0,
-                    0.0, 0.06283202,0.2111851, 0.0,
-                    0.0, 0.47909424,0.6283214, 0.0,
-                    0.0, 0.0,       0.0,       0.0
-                ]
-            );
-
-            let mut dem_for_big_scale = make_dem(&kernel::tests::dems::bigger_dem());
-            let temp_db_for_big_scale = tempfile::NamedTempFile::new().unwrap();
-            let compute_big_scale = super::compute(
-                &mut dem_for_big_scale,
-                super::Config {
-                    dem_metadata: crate::cpu::storage::metadata::MetaData {
-                        scale: 100.0,
-                        ..crate::run::compute::test::big_dem_metadata()
-                    },
-                    ..super::default_config(Backend::CPU, &temp_db_for_big_scale)
-                },
-            );
-            #[rustfmt::skip]
-            expect_eq!(
-                compute_big_scale.total_surfaces,
-                [
-                    0.0, 0.0,      0.0,       0.0,
-                    0.0, 628.317,  2111.8496, 0.0,
-                    0.0, 4790.928, 6283.1714, 0.0,
-                    0.0, 0.0,      0.0,       0.0
-                ]
-            );
-        }
-    }
-}
