+++ crates/total-viewsheds/src/storage/db.rs
