+++ crates/total-viewsheds/src/storage/engine.rs
